Brach

+++ core/src/main/java/com/github/chat/network/WebsocketRoomMap.java
@@ -8,11 +8,11 @@ import java.util.Map;
 
 public class WebsocketRoomMap {
 
-    private Map<Long, WebsocketConnectionPool> roomSession;
+    private Map<Integer, WebsocketConnectionPool> roomSession;
 
     private WebsocketConnectionPool websocketConnectionPool;
 
-    public void addSession(Long idRoom, String nickname, Session session) {
+    public void addSession(Integer idRoom, String nickname, Session session) {
         if (this.roomSession.containsValue(idRoom)) {
             this.roomSession.get(idRoom).addSession(nickname, session);
         } else {
@@ -22,16 +22,16 @@ public class WebsocketRoomMap {
         }
     }
 
-    public List<Session> getSessions(Long idRoom) {
+    public List<Session> getSessions(Integer idRoom) {
         WebsocketConnectionPool room = this.roomSession.get(idRoom);
         return room.getSessions();
     }
 
-    public void removeUserFromSession(Long idRoom, String nickname) {
+    public void removeUserFromSession(Integer idRoom, String nickname) {
         this.roomSession.get(idRoom).removeSession(nickname);
     }
 
-    public void closeSession(Long idRoom, String nickname){
+    public void closeSession(Integer idRoom, String nickname){
         try {
             this.roomSession.get(idRoom).getSession(nickname).close();
         } catch (IOException e) {
